# Fix unyt_array.__getitem__ to support numpy masked arrays