--- a/unyt/array.py
+++ b/unyt/array.py
@@ -1624,14 +1624,11 @@ class unyt_array(np.ndarray):
 
     def __getitem__(self, item):
         ret = super(unyt_array, self).__getitem__(item)
-        if ret.shape == ():
-            return unyt_quantity(
-                ret, self.units, bypass_validation=True, name=self.name
-            )
-        else:
-            if hasattr(self, "units"):
-                ret.units = self.units
-            return ret
+        if hasattr(ret, "shape") and ret.shape == ():
+            ret = unyt_quantity(ret, self.units, bypass_validation=True, name=self.name)
+        elif hasattr(ret, "units"):
+            ret.units = self.units
+        return ret
 
     #
     # Start operation methods