--- a/unyt/array.py
+++ b/unyt/array.py
@@ -1624,10 +1624,12 @@ class unyt_array(np.ndarray):
 
     def __getitem__(self, item):
         ret = super(unyt_array, self).__getitem__(item)
-        if hasattr(ret, "shape") and ret.shape == ():
-            ret = unyt_quantity(ret, self.units, bypass_validation=True, name=self.name)
-        elif hasattr(ret, "units"):
-            ret.units = self.units
+        if getattr(ret, "shape", None) == ():
+            ret = unyt_quantity(ret, bypass_validation=True, name=self.name)
+        try:
+            setattr(ret, "units", self.units)
+        except AttributeError:
+            pass
         return ret
 
     #