Closes #169