@l-johnston 2020-08-24
@ngoldbaum I ran the `pyperf` [benchmark](https://gist.github.com/l-johnston/75293a7b072d37c6629c58de8a36fce5) and get the following results:

- v2.7.2: 2.83 µs (SD=0.11 µs)
- proposed: 2.97 µs (SD=0.16 µs)