@jzuhone 2021-08-17
It would be very nice to figure out the performance impact first.