@neutrinoceros 2021-08-17
@l-johnston thanks. I'm now running a modified version of you benchmark with a much heavier array