@neutrinoceros 2021-08-17
So I ran your benchmark with a 3GiB array, each run took 10min on my machine and this is what I'm getting out of it:


current master: 
Mean +- std dev: 1.26 us +- 0.09 us

this patch:
Mean +- std dev: 1.26 us +- 0.17 us

so the standard deviation is definitely affected, but the raw performance not so much, as far as I can tell.